% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_margins.R
\name{getDifferencesRelative_noMod}
\alias{getDifferencesRelative_noMod}
\title{Get Relative Differences for Binary Exposure}
\usage{
getDifferencesRelative_noMod(results, binary_exposure)
}
\arguments{
\item{results}{A model object, typically from `marginaleffects`, containing the results of the model.}

\item{binary_exposure}{A string representing the binary exposure variable for the analysis.}
}
\value{
A dataframe with relative differences, including the estimate, confidence intervals, and p-value.
}
\description{
This helper function calculates relative differences for a binary exposure variable using the `marginaleffects` package.
}
\examples{
df <- mtcars
results <- lm(mpg ~ vs, data = df)
getDifferencesRelative_noMod(results, binary_exposure = "vs")
}
\keyword{internal}
