% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab1.R
\name{tab1}
\alias{tab1}
\title{Prints frequency tables for specified variables in a data frame.}
\usage{
tab1(df, vars)
}
\arguments{
\item{df}{A dataframe or tibble containing the variables to tabulate.}

\item{vars}{A character vector specifying the names of the variables in the data frame.}
}
\value{
This function prints the frequency tables for the specified variables.
}
\description{
Uses approach that is similar to Stata's 'tab1'
}
\examples{
tab1(mtcars, "mpg") # Use for one variable.
tab1(mtcars, c("mpg", "cyl", "disp")) # Use a character vector for multiple.
}
